Make BufferedSender an interface and Hide the Constructor
