+++ sdk/search/search-documents/src/searchClient.ts
@@ -48,7 +48,8 @@ import { odataMetadataPolicy } from "./odataMetadataPolicy";
 import { IndexDocumentsBatch } from "./indexDocumentsBatch";
 import { encode, decode } from "./base64";
 import * as utils from "./serviceUtils";
-import { SearchIndexingBufferedSender } from ".";
+import { SearchIndexingBufferedSender } from "./searchIndexingBufferedSender";
+import { createSearchIndexingBufferedSender } from "./searchIndexingBufferedSenderImpl";
 
 /**
  * Client options used to configure Cognitive Search API requests.
@@ -627,7 +628,7 @@ export class SearchClient<T> {
   public getSearchIndexingBufferedSenderInstance(
     options: SearchIndexingBufferedSenderOptions = {}
   ): SearchIndexingBufferedSender<T> {
-    return new SearchIndexingBufferedSender(this, options);
+    return createSearchIndexingBufferedSender(this, options);
   }
 
   private encodeContinuationToken(
